backend/requirements.txt
Django==5.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
firebase-admin==6.2.0
python-decouple==3.8
